+++ R/globals.R
@@ -1,0 +1,1 @@
+utils::globalVariables(c("model", "year", "x", "y", "cellID", "val"))